+++ src/domain/news/feed/model.py
